cros-board: add a prune-list to the eclass
Prevent new users of this eclass by comparing to a prune-list.  New
packages should not be added to this list.

This list was generated by sourcing all ebuilds in an internal
checkout and searching for the ALL_BOARDS variable.  If it was
defined at the time this CL was authored, it was added to the list.

BUG=b:194831262
TEST=manually, add a package which uses cros-board not on this list,
     observe the error
TEST=CL:3088811 dry run (forces complete build)

Disallow-Recycled-Builds: test-failures
Change-Id: I90ccf9b2d41603bff4d65bf6dca51ac9ecf8900b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3083147
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>