ambassador: Add moonbuggy coreboot configuration
Create a new coreboot configuration for the moonbuggy variant
of the ambassador reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:191356135
TEST=FW_NAME=moonbuggy emerge-ambassador coreboot chromeos-bootimage
Ensure that image-moonbuggy.*.bin are created

Change-Id: I090eb5a15be56ebe7e36e19d1acdb3a074832295
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3107169
Reviewed-by: Jeff Chase <jnchase@google.com>
Tested-by: Rehan Ghori <rehang@chromium.org>
Commit-Queue: Rehan Ghori <rehang@chromium.org>