dev-rust/as-slice: correctly depend on multiple versions of generic-array
The as-slice crate intentionally depends on all three versions of the
generic-array crate. We need all three to be present in the local cargo
registry when compiling. But the way cargo2ebuild.py had encoded the
version constraints does not work correctly in this very unusual case.
Portage would resolve the dependencies by install 0.12.4 and 0.14.4 but
not 0.13.3, because:

0.12.4 < 0.13.0
0.12.4 < 0.14.0
0.12.4 < 0.15.0
0.14.4 >= 0.12.4
0.14.4 >= 0.13.3
0.14.4 >= 0.14.4

We can work around it by depending on the exact version and slot of all
three versions of generic-array.

BUG=b:191820715
TEST=emerge dev-rust/as-slice

Change-Id: Ie6f38deea91f384e3562d58e256072aee9c5e175
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3105857
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>