media-gfx/airscan: Stop setting fuzzer options directly
Currently this ebuild uses fuzzer-setup-binary to configure the build
for fuzzing. This is unnecessary and incorrect.

To build for fuzzing you should add:
-fsanitizer={address,memory,undefined} to run with sanitization
-fsanitizer=fuzzer-no-link for things that are not fuzzer executables
-fsanitizer=fuzzer *only* when linking the actual fuzzer binary

The fuzzer profiles add the first two globally already, leaving it up
to individual ebuilds to add the second where appropriate. OTOH the
functions above add all three flags, which is only appropriate if the
ebuild is *only* building fuzzer binaries.

Historically that happened to be the case, but adding an exe_wrapper
to meson causes it to build a (non-fuzzer) test binary to validate
that it really can produce runnable binaries.

The meson build file for this project already adds -fsanitizer=fuzzer
to the appropriate executables, and not to anything else, so we can
actually just get rid of this call and use the default options.

We still need sanitizers-setup-env because the non-fuzzer sanitizer
profiles work by setting USE flags instead which need to be read by
that function.

BUG=chromium:1227738
BUG=chromium:1239482
TEST=emerge-amd64-generic media-gfx/airscan

Change-Id: I0891280c5cdb3a9ad776ee3ab190946ca8cd4cbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3106627
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>