eclass/cros-rust: split up src_configure function
The cros-rust_src_configure function currently assumes the ebuild is for
a single Rust crate with Cargo.toml at the root. But we are working on
packaging a project with multiple interdependent crates in
subdirectories (as well as some non-Rust code as well).

Most of the work done in the cros-rust_src_configure function is setting
up a cargo configuration file, and exporting various environment
variables to go with it. This needs to be done only once, and is
applicable to all crates in the ebuild.

However, it also tries to regenerate or remove Cargo.lock, presumably so
that cargo will not complain if the dependency versions in our local
registry are different to the ones specified in Cargo.lock. This step
assumes that Cargo.toml and Cargo.lock are in the root of the source
tree.

Split cros-rust_src_configure into two separate helper functions, so
that the ebuild can set up cargo configuration once, but regenerate
Cargo.lock in multiple subdirectories if needed (or not at all, in case
there is no Cargo.lock committed).

BUG=b:191820715
TEST=emerge hps-firmware

Cq-Depend: chromium:3113367
Change-Id: I5c09356c28e7e857739e6f6138116084a7819d82
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3106629
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Tested-by: Dan Callaghan <dcallagh@chromium.org>