gobject-introspection: Support cross compilation
Significant changes to the ebuild in order to make cross compilation
work in the Chromium OS build.

Add wrappers for `ldd` and cross-compiled introspection binary execution.
The purpose and behavior of these wrappers are explained in the header
comments in those files.

Add `cros_host` USE flag. This flag is only set when building for the
build host, so use the absence of this USE flag to key when we need to
BDEPEND on gobject-introspection and QEMU. The gi_cross_use_prebuilt_gi
configuration option allows us to use the build host's g-ir-compiler
during the build instead of trying to execute a cross-compiled binary.
This setting, along with the setting to make use of the wrapper scripts,
is also gated behind the absence of the `cros_host` USE flag.

The PKG_CONFIG variable for the build target needs to be exported in
both src_configure and src_compile otherwise the build will be poisoned
with include paths that are not safe for cross-compilation.

This CL also requires the CL in Cq-Depend to have landed and for the
prebuilts for those boards to have been regenerated, otherwise the
introspection binaries produced during this package's build fail due to
encountering instructions not supported on the build host and that QEMU
is not capable of emulating.

BUG=b:188712021, chromium:1099508
TEST=`emerge-grunt gobject-introspection`
TEST=`emerge-elm gobject-introspection`

Cq-Depend: chromium:3111209
Change-Id: I5a1d2e01dbfc4a69f6934f96be5a33580a1c47ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3111450
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>