libqrtr-glib: Disable introspection during build
Introspection requires executing compiled binaries for the target, which
is not typically possible when cross-compiling. This behavior has a
default setting of 'auto', so explicitly set it to 'no' now that
gobject-introspection is cross-compiling so that this package doesn't
try to execute any cross-compiled binaries during its build.

BUG=b:188712021
TEST=CQ

Change-Id: Ifc27c17342938426a44f5a7faaa6baf883aa0156
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3112637
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>