chromeos-config: move dependency on bsp packages to eclass
Since a subslot dependency is non-transitive[1], packages which depend
on chromeos-base/chromeos-config won't rebuild when config changes
happen.

With the new non-virtual dependency on the bsp backed by USE flags, we
can encode this dependency directly into the eclass, which most
packages inherit in order to use cros_config_host.

Then, we can add an environment variable (CROS_UNIBUILD_ECLASS)
exported by this eclass as well.  The purpose of this is so that we
can later add support to cros_config_host to validate that the eclass
was inherited, ensuring that all packages that use chromeos-config
have properly specified dependencies.

[1]: If A has a subslot dependency on B, and B has a subslot
dependency on C, and package C changes, B will be rebuilt but not A,
which is quite an unfortunate feature of portage, IMO.

BUG=b:197038156
TEST=emerge-guybrush chromeos-config

Change-Id: Ia8c0ce9cefa335cdb2c908edccc5374bcf5938c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3115054
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>