F20: Add gtest metadata generation to test image
In order to use gtest as a first class harness, a system needs to be in
place to generate test metadata for the tests. The tests will be spread
through the source tree, this setup allows for a single location to define
metadata for gtest tests without requiring pulling down the entire
source tree.

BUG=b:198202240
TEST=emerge gtest-tests and emerge autotest-all

Cq-Depend: chromium:3123329, chromium:3139178
Change-Id: If136864bf6045f34f542bcd94cf23b84b8955049
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3116544
Tested-by: Jesse McGuire <jessemcguire@google.com>
Commit-Queue: Jesse McGuire <jessemcguire@google.com>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>