Uprev protofiles to support only one additional policy
This is not a regular uprev. The normal uprev could not be
performed due to a lot of errors in Authpolicy occur when
picking up the new proto changes.

This uprev is mainly adding support to one additional policy proto:
1) DevicePciPeripheralDataAccessEnabledProtoV2

Its important to notice that the pulled policy's patch has
been modified manually to remove DeviceScheduledReboot policy
from being pulled out along the way. That happened because
of the following:
1) We don't need that policy for that workaround; The failures
that exit (and linked in the bugs) happened because the
inexitance of that policy proto encode/decode in Authpolicy
2) The DeviceScheduledReboot won't be pulled fully with the
*pci-*v2.patch. Only its proto definition will be pulled, but
won't be able to encode/decode/test it out which seemed
as worse workaround than changing the contents of the patch

This policy was crucial and without its usages, some
boards were failing for specific use cases. See the attached
bugs for more information.

There will be additional two dep CLs to fix the dependency of the
Authpolicy to >= *48 version of protofiles and adding the
corresponding encoding for each one of them.

BUG=chromium:1234804, b/193441567
TEST=./build_packages --board=amd64-generic && \
     cros_run_unit_tests --board=amd64-generic --packages authpolicy

Cq-Depend: chromium:3118374, chromium:3118231
Change-Id: I75b754a7acaf3744de23d464dcdb0373c49b32ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3118372
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Felipe Andrade <fsandrade@chromium.org>
Tested-by: Amr Aboelkher <amraboelkher@chromium.org>
Commit-Queue: Amr Aboelkher <amraboelkher@chromium.org>