chromeos-base/policy-go-proto: Update protofiles dependency
Due to recent issues with the latest uprev, and the crucial need
to have a specific policy to be picked up, there was a workaround
that has been done to retrieve its specific proto changes patch
and use it until the original uprev to be done.

Given that DevicePciPeripheralDataAccessEnabledProtoV2 policy
is needed to be used in tast (Go code), it will be needed
to reflect that in the protofiles version dependency.

That CL should be merged after crrev.com/c/3118372 has been
merged successfully.

This CL is changing the dependency of the protofiles to the latest
one that contain the workaround patch of that needed policy.

BUG=chromium:1234804, b/193441567
TEST=emerge-eve policy-go-proto

Change-Id: Ic45cdc03a7359f1350f71b2db0117b134b49d71c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3122330
Tested-by: Amr Aboelkher <amraboelkher@chromium.org>
Commit-Queue: Amr Aboelkher <amraboelkher@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Felipe Andrade <fsandrade@chromium.org>