brya: Add anahera coreboot configuration
Create a new coreboot configuration for the anahera variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:197850509
TEST=FW_NAME=anahera emerge-brya coreboot chromeos-bootimage
Ensure that image-anahera.*.bin are created

Change-Id: Ic654300400f05d341d87a3bccda6724e17786c6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3121934
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Tested-by: Wisley Chen <wisley.chen@quanta.corp-partner.google.com>
Commit-Queue: Tim Wawrzynczak <twawrzynczak@chromium.org>