net-analyzer/wireshark: upgrade 1.10.5 -> 3.4.7
This is quite a massive jump for wireshark due to v1.10.5
breaking against newer glib2 and being obsolete.

This minimizes differences with upstream Gentoo but they
are enough to keep this ebuild away from portage-stable
for now in the interest of unblocking the glib2 upgrade.

The main differences from gentoo upstream are 2 patches
fixing cross-compile bugs (should be sent to wireshark
upstream), a lua eclass dependency and some pcap user
management ebuild which do not exist in CrOS.

BUG=b:197671366
TEST=Local builds on arm, arm64, amd64; CQ.

Change-Id: I4b7196e877df26d478197f2203e051c6636b9931
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3122333
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>