dev-util/perf: Control use of zstd via a USE flag
By default perf will look for the presense of libzstd in the sysroot
and use that to control this feature. This is bad because it
introduces a runtime dependency portage doesn't know about, so zstd
may not be installed when perf is.

Fix this by adding a new USE flag that controls this behaviour. If
set, zstd is added to the dependency list (and will be installed),
otherwise zstd support is disabled.

BUG=chromium:1244439
TEST=emerge-amd64-generic perf with zstd already installed
TEST=readelf -d /build/amd64-generic/usr/bin/perf doesn't list libzstd

Change-Id: I2b3764e211e86b15f34e7135972b4204cd9646a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3128402
Tested-by: Fergus Dall <sidereal@google.com>
Auto-Submit: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Commit-Queue: Sergey Frolov <sfrolov@google.com>
Reviewed-by: Sergey Frolov <sfrolov@google.com>