fuse: Import fuse 3
Import libfuse3 alongside libfuse2 as a separate slot (3).
libfuse2 remains slot 0 and thus the default.
libfuse3 will be needed for drivefs soon.

libfuse2 and 3 are designed to both run alongside each
other which is why they are packaged as separate slots.

libfuse3 is not used by any other package yet and this
version is also not ready to be used yet, as we still
need to apply patches in a follow-up.

BUG=chromium:1242814
TEST=emerge-zork sys-fs/fuse:3, cros deploy sys-fs/fuse-3.10.4

Change-Id: I4715959126dbfbe5abea299f133aba4f599987e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3127962
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Tested-by: Peter Marshall <petermarshall@chromium.org>