brya: Add brask coreboot configuration
Create a new coreboot configuration for the brask variant
of the brya reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:196941353
TEST=FW_NAME=brask emerge-brask coreboot chromeos-bootimage
Ensure that image-brask.*.bin are created

Change-Id: I31161c700a987574c34a652fec31daa8e11c939d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3139963
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>