tensorflow: Slotted dep on flatbuffers in 1.14.0
Change the tensorflow 1.14.0 version's ebuild to use a slotted
dependency on flatbuffers.

This will help in upreving the flatbuffers to 2.0.0, since with the
major version update also the library SONAME gets incremented (to
"libflatbuffers.so.2"). Without a slot dependency, packages that
use tensorflow 1.14.0 will flakily fail the build, e.g., if prebuilts
are enabled and a version that's linked against libflatbuffers.so.1 is
fetched. A follow-up CL will change the flatbuffers' ebuild to specify a
subslot in order to trigger rebuilds in cases like this.

Also the Portage doc suggests that the slot dependency should've been
used here:

  Generally, consumers which link against a library possessing a subslot
  that represents SONAME or binary compatibility should subscribe to it
  (request to be rebuilt when the subslot changes) with :=.
  https://devmanual.gentoo.org/general-concepts/slotting/index.html

BUG=b:197500482
TEST=uninstall flatbuffers and do emerge-$BOARD tensorflow=1.14.0

Change-Id: Ib67b6e9cedaa1cea58ed9603bad38748ada6f366
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3141476
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>