gobject-introspection: Hardcode GOBJECT_INTROSPECTION_LIBDIR path
Define the baked-in value of GOBJECT_INTROSPECTION_LIBDIR as the path
that will exist on a test image, so that the search path is correct even
without the ability to specify GI_TYPELIB_PATH through the environment.

Also install an env.d file specifying both paths so that anything that
runs and picks up settings from `/etc/profile` will find the typelib
files regardless of where they are installed.

BUG=b:188712021
TEST=None

Change-Id: I08d27de5ffc592da0956d5c0489043c7bb2fa31a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3146572
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>