aosp-frameworks-ml-nn: remove /etc/env.d/drivers file
This file was set up for multi-HAL, but we're not using that at the
moment and it interferes with VTS / CTS in tast. We want to move that
functionality to chromeos-config anyway, so remove this file for now.

BUG=b:199989232
TEST=tast run <DUT> platform.NNAPIDriver.*

Change-Id: I4a53114bfe92d26e09da5ef6a5bd16ab10a1d73f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3161669
Tested-by: Jim Pollock <jmpollock@chromium.org>
Auto-Submit: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>