disable test phase in Rust ebuilds where it is currently broken
The cros-rust.eclass defines src_test() to run `cargo test`. But that
fails in a large number of our Rust crate ebuilds currently.
In most cases, it seems to be because dev-dependencies are being
(intentionally) excluded but the crate needs those to compile its tests.

BUG=chromium:1242244
TEST=emerge -C dev-rust/\* ; FEATURES=test emerge hps-firmware prjoxide

Change-Id: I201116afb90a3873fa180ef687983b39832cd00a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3168463
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>