perfetto: open producer socket access to all users
This changes the permission mode bits of the producer socket
(/run/perfetto/traced-producer.sock) from 0660 to 0666 to allow all
users. This is done to make it easy for new producers to send trace data
without the need for adding the user to group 'traced-producer' and
also update chromite user/group tests.

Security-wise: the producer socket is designed to be unprivileged.
traced doesn't allow readback of trace data from the socket. Security is
ensured by fuzzing and minijail sandboxing of traced.

BUG=b:199891591
TEST=CQ dry run.

Change-Id: I8f0462356fa5465a52ec30d702e656b3cd9dc3f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3185721
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Tested-by: Chinglin Yu <chinglinyu@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>