chromeos-base/hps-firmware/: objcopy to flatten elf fw images
HPS deals with flat binaries and not ELF's directly, therefore
flatten the ELF's to flat binaries ready to be programmed. This
is also required to allow CQ not to try and parse arm32-little
on x86.

There is a issue where the linker GC the image to a 0B one
however this unblocks firmware building within the chroot
and as part of CQ. The linker issue will be resolved in
a follow up.

BUG=b:198099657
TEST=`emerge-brya hps-firmware`.

Change-Id: Ia6e471e79ec154e6da5f5f4b6b70b1d715a52f59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3188772
Commit-Queue: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Auto-Submit: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Dan Callaghan <dcallagh@chromium.org>