chromeos-base/hps-firmware: drop unnecessary target config
When the ebuild runs cargo, it is already obeying both
${ECARGO_HOME}/config and .cargo/config from the hps-firmware source
tree. We don't need to repeat these rustflags here. If we do, we will
end up with the linker scripts passed twice.

This was not a problem before, because rustflags in the cargo
configuration were being ignored anyway. When we fix that, the duplicate
linker scripts become a problem.

BUG=b:201504188
TEST=emerge hps-firmware

Cq-Depend: chromium:3195158
Change-Id: I94d936690c11a3486f20de29b65396351e2310bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3195356
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Tested-by: Edward O'Callaghan <quasisec@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>