sys-kernel/linux-headers: add upstream V4L2 VP8 stateless controls
Add the official stable VP8 stateless controls so Chromium can use them.
The controls are up-to-date as of commit e6a7d7c342cb, the last commit
to touch the VP8 stateless controls before the v5.13 release.

This simply makes the control definitions available to userspace. It is
up to the individual userspace components to switch over.

BUG=b:198118193
TEST=emerge sys-kernel/linux-headers and check for
     /usr/include/linux/media/vp8-ctrls-upstream.h

Change-Id: I6f58e117c52e8e1d0b2553f19b8635a5b6935502
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3199902
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Chen-Yu Tsai <wenst@chromium.org>
Commit-Queue: Chen-Yu Tsai <wenst@chromium.org>