samba: Install smbd to /usr/local/sbin to provide test server
The test infrastructure at the moment relies on a Samba instance inside
Crostini to test. This introduces a whole host of problems not related
to SMB and makes it difficult to promote the test to stable. Having smbd
available directly on the test image ensures tests are much stabler.

NOTE: The /usr/local/* directory is only present on test images so this
file is not available on release images. I've left a comment clarifying
as such.

BUG=chromium:1156844
TEST=emerge-hatch samba && cros deploy kled samba
TEST=./usr/local/sbin/samba --interactive

Cros-Experiments: chromeos.cros_test_platform.add_resultdb_settings
Change-Id: I74f9908b1ff7fbf20c9257d25a6de17fa0d0b77f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3222327
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>