zephyr: Separate out PoC vs. normal builds
For Zephyr proof-of-concept builders, we intentionally disable all
models without a Zephyr EC at the cros_config level.  However, for
real builders, we don't want to enable this condition.

This adds a new USE flag, zephyr_poc, which is enabled for
{asurada,trogdor,volteer}-zephyr only.  Only for these builds will we
strip out non-Zephyr models.  For herobrine, we'll simply not pack the
EC for hoglin/piglin, which use a legacy CrOS EC with non-upstreamed
patches from Qualcomm.

This requires adding logic to the chromeos-bootimage ebuild to not
pack the EC for devices which don't have an EC build target, since
we've never supported some devices with an EC and some without in the
same overlay before.

BUG=b:201730475
TEST=build_packages --board herobrine
     inspect firmware outputs, notice builds for hoglin/piglin
     herobrine has Zephyr EC in CBFS

Cq-Depend: chrome-internal:4187332,chrome-internal:4187334,chrome-internal:4187335
Change-Id: I8087df154855ffee307d4d3693c1f6c4245ccdf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3223564
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Bob Moragues <moragues@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>