cros-rust.eclass: quote rustflags elements to avoid tricium complaint
Tricium says "warning: Use spaces, not commas, to separate array elements"
because we have an array element with a comma in it. This shows up
as a warning on every CL that affects cros-rust.eclass. This CL
quotes the array elements to avoid the warning.

BUG=None
TEST=See that Tricium does not complain anymore

Change-Id: I5eeebb8ad0e6ae604d7aee1b2225f673ea3f24ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3227715
Tested-by: Bob Haarman <inglorion@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>