ntfs3g: Don't log volume label name
The volume label represents PII and should not be captured in logs. Add
a patch that removes the label from both the mounting and unmounting log
event. This happens in lowntfs-3g as well but this is removed on
ChromeOS so don't apply this via the patch.

BUG=chromium:1240059
TEST=emerge-hatch ntfs3g, cros deploy KLED ntfs3g
TEST=Plug device, observe /var/log/messages, unmount and observe as well

Change-Id: I204efff537e0a169e68f1dc6981775265590b894
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3229245
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>