dev-go/syzkaller: explicitely set GOPATH mode
We plan on upgrading Go 1.15 -> 1.17 which will change
the default GOPATH to modules mode, but until all go
packages are converted, the old GOPATH mode will be
explicitely enable via the eclass.

syzkaller is the only exception which overrides the
eclass compile function so we need to set the var in
the ebuild until modules migration, at which point
the variable will be dropped.

BUG=b:187795288
TEST=Local builds on arm, arm64, amd64, bootstrapped sdk, CQ.

Change-Id: If53873ba0bc493eb1c8b1410a41b8533155ca1d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3238290
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>