net-print/cups: Tie fuzzer builds to debug builds
This change modifies the way in which CUPS is built under the fuzzer
profile. We would like CUPS to build s.t. line numbers are visible in
stack traces. It appears to suffice to build CUPS with `USE=debug` to
achieve this.

This CL makes no attempt to adjust the optimization level in the
presence of `USE=debug`; in testing, adding `--with-optim=-Og` to the
`configure` invocation appears to have the _undesirable_ effect of
obscuring line numbers from the stack trace. (Additionally, `-Og` ends
up intermingling with the existing `-Os` set by default, which is
probably a bug we need to root out.)

BUG=b:204370124
TEST=./setup_board ... --profile=fuzzer && cros_fuzz reproduce ...

Change-Id: Ifbed0e82431bbb68adeb30e7d26d611a89d81875
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3248838
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Kalvin Lee <kdlee@chromium.org>
Commit-Queue: Kalvin Lee <kdlee@chromium.org>