dhcpcd: Stop only active interfaces
When stopping interfaces via DBus we should check if the interface is
active before calling `stop_interface()`.  This is because only active
interfaces have been configured - otherwise we could end up accessing
fields that have not been initialized.

Concrete example is when shill crashes - then dhcpcd is notified via
DBus and we are stopping all interfaces.  However one of them is
loopback (which is left unconfigured) and any attempt to stop it ends up
with a crash.

BUG=b:204389871
TEST=emerge-${BOARD} dhcpcd, deploy, kill shill (-9) and check in logs
     and /var/spool/crash that dhcpcd is not crashing.

Change-Id: I2dcb01e1182fe321cd200f697c23098cfd312e94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3257271
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>