Reland "featured: Add a library to check feature state."
This is a reland of 24c6fbe4bc3982deb97496404892e4ba6b80fe13

The linked change fixes the bug in the original CL, which was caused by
a test that verified that we crashed when using distinct feature structs
with the same name. However, that check was a DCHECK, so we should only
expect to crash if DCHECKs are on.

Original change's description:
> featured: Add a library to check feature state.
>
> BUG=b:203674129
> TEST=unit tests
>
> Cq-Depend: chromium:3260651
> Change-Id: I65b6e2ffb0820b7100ff07d1656a7304eb7f7322
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3261097
> Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
> Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
> Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>

BUG=b:203674129
TEST=unit tests

Cq-Depend: chromium:3283272
Change-Id: I51ade0654fa23639d487ca6ba966341ffaccc647
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3283271
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>