libchrome: add IUSE flag board_use_mistral
A backward compatibility patch will be added in r896220 uprev to include
an extra base/logging.h.
This is only added as a temporary measure to unblock uprev since Gwifi
boards is planning on only using release R94 branch and is not accepting
any change until they resolve the branching issue.
Add this IUSE flag so to apply the patch selectively only on mistral.

BUG=b:198993705
TEST=emerge-mistral libchrome at r892660 with unmodified webserver

Change-Id: Ia69e8486ee9d1411caad1d63be1e852c0d01747a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3261997
Tested-by: Grace Cham <hscham@chromium.org>
Reviewed-by: Qijiang Fan <fqj@google.com>
Commit-Queue: Qijiang Fan <fqj@google.com>