dedede: Add beadrix coreboot configuration
Create a new coreboot configuration for the beadrix variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:204882915
TEST=FW_NAME=beadrix emerge-dedede coreboot chromeos-bootimage
Ensure that image-beadrix.*.bin are created

Cq-Depend: chromium:3319030
Change-Id: Ic58239f4eb9ded816ec90d715a62b263b47d8622
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3284344
Reviewed-by: Teddy Shih <teddyshih@ami.corp-partner.google.com>
Tested-by: Teddy Shih <teddyshih@ami.corp-partner.google.com>
Reviewed-by: Henry Sun <henrysun@google.com>
Commit-Queue: Henry Sun <henrysun@google.com>