chromeos-cr50-dev: Remove dependency on EC build target name
The cr50 ebuild has an odd dependency on the EC build target name to
just enable extra binaries for the reef board.  I'm not sure the
historical context as to why reef is the only board that needs these
binaries, but the inherit of cros-ec-board needs to go so that we can
assume unibuild only there.

BUG=b:187699715
TEST=emerge-reef chromeos-cr50-dev

Change-Id: I26ab35566812665d356126955c9411fb353c318d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3276363
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>