target-chromium-os-test: Drop cros_ec USE flag for ec-devutils
The purpose of installing ec-devutils is solely so that one chromebook
with a test image can be used to debug/flash another in a chain.  It
should not depend whether we are currently building the EC on ToT or
not for this dependency to get activated.

BUG=none
TEST=CQ

Change-Id: Idde1a44bab17ab9deaa175c154d91d67bfae8130
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3276364
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>