brya: Add vell coreboot configuration
Create a new coreboot configuration for the vell variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:205908918
TEST=FW_NAME=vell emerge-brya coreboot chromeos-bootimage
Ensure that image-vell.*.bin are created

Change-Id: Iaf625da858dc1b20fbb5f3e983f7792141027333
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3277704
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>
Tested-by: Shon Wang <shon.wang@quanta.corp-partner.google.com>
Commit-Queue: YH Lin <yueherngl@chromium.org>