chromeos-base/policy-testserver: Support python 3 and EAPI 7
This ebuild is still using the old python version, along with the old
EAPI as well.

Depot_tools are supporting now the newest python version (i.e. v3).

This CL is migrating the python version to 3.0, and the EAPI to the latest which is  "7".

BUG=none
TEST=CQ

Cq-Depend: chromium:3279470
Change-Id: I2bc6d20615571832556e4df81aa43a15c9307284
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3282153
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>