cros-kernel2: Use same toolchain for module splitdebug stripping
If we build the kernel with clang, then we should use the same toolchain
to run split and objcopy so that we're consistent. Unfortunately, the
SPLIT environment variable is overwritten by the time we get into
strip_splitdebug so let's add another STRIP variable, REAL_STRIP, that
we can use in the script to call the STRIP that is intended, either
llvm-strip or GNU's version. The OBJCOPY variable isn't overwritten so
we can simply use that instead of CROSS_COMPILE.

Keeping this consistent fixes a problem where the kernel is compiled
with clang and is software tagging high bits of pointers with
KASAN_SW_TAGS and binutils strip doesn't understand that.

BUG=None
TEST=emerge-${BOARD} chromeos-kernel-${X}_${Y}

Change-Id: I56b3f8e6c06a112f18658d2f0e9c7838503d5223
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3290672
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>