REVERT-ME: modemmanager-next: Use antenna index 0 for dynamic sar
As per spec, setting the antenna index to 0xFFFFFFFF
should set the backoff index to all antennas on the system.
This feature is currently not supported in
Fibocom FW 18500.5001.30.05.27.12.
This will be fixed in the next MR.
Please revert this patch once the new FW lands.

BUG=b:188002987
TEST=emerge-zork modemmanager-next && cros deploy
     mmcli -m 0 --sar-set-power-level=0

Cq-Depend: chromium:3290576
Change-Id: I6d1ecde6a5c1da47ec4948def3912d453b44477a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3290675
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Madhav <madhavadas@google.com>
Commit-Queue: Madhav <madhavadas@google.com>