mesa-reven: sync the DEPEND vars between 9999 and stable
The mesa-reven ebuilds were copied from the mesa ebuilds which set
CROS_WORKON_MANUAL_UPREV, so they get out of sync easily. In the mesa
9999 ebuild, the DEPEND variables were reworked with a COMMON_DEPEND
variable and a BDEPEND, but these changes were never propagated to the
stable ebuild.

For mesa-reven we want to eventually turn off CROS_WORKON_MANUAL_UPREV,
so it would be good to get the ebuilds more in sync to make that
transition easier. Sync the stable ebuild's DEPEND vars to match the
9999 version, and add the new libglvnd IUSE.

One edit to the 9999 version to make it more like the stable
version: the dep on sys-devel/llvm should only be in DEPEND, not
RDEPEND.

BUG=b:200806429
TEST=Check that the DEPEND lines match in the ebuilds with "diff -u mesa-reven-9999.ebuild mesa-reven-21.0.ebuild"
TEST=emerge-reven mesa-reven

Change-Id: I6bbdb109327474d279dec2bb6764a79f4b4fdb44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3291538
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Michael Pobega <pobega@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>