brya: Add taniks coreboot configuration
Create a new coreboot configuration for the taniks variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:207402720
TEST=FW_NAME=taniks emerge-brya coreboot chromeos-bootimage
Ensure that image-taniks.*.bin are created

Cq-Depend:chromium:3319025
Change-Id: Id9948847d205f169d42bfaaac3ac43760c8af63e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3301939
Reviewed-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: Parth Malkan <parthmalkan@google.com>
Tested-by: Joey Peng <joey.peng@lcfc.corp-partner.google.com>
Commit-Queue: Joey Peng <joey.peng@lcfc.corp-partner.google.com>