app-emulation/lxd: explicitly enable Go modules mode
Starting with PR [1] and [2] lxd supports Go modules mode
which is also the default since Go 1.17 and does not work
with the legacy GOPATH mode anymore.

This is not a problem for distributions like Gentoo which
support Go >= 1.17 with modules only, but CrOS still uses
Go 1.16 with GOPATH and is in a transition period to 1.17
with modules.

Until the transition is complete, packages like lxd which
do not work with GOPATH anymore will need to explicitly
enable modules mode.

After CrOS finishes the transition to modules by default,
this line can be dropped because it will be a no-op.

[1] https://github.com/lxc/lxd/pull/9138
[2] https://github.com/lxc/lxd/pull/9142

BUG=b:187795288
TEST=Local builds on x86_64, arm, arm64; CQ

Change-Id: Iced4e7b6100b26581fe79baa4350fdc6cac1f519
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3306397
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Auto-Submit: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>